+++ tests/testthat/test_checkSubset.R
@@ -10,6 +10,8 @@ test_that("checkSubset", {
   expect_true(testSubset(character(0), letters, empty.ok = TRUE))
   expect_false(testSubset(NULL, letters, empty.ok = FALSE))
   expect_true(testSubset(character(0), letters, empty.ok = TRUE))
+  expect_false(testSubset(NULL, letters, empty.ok = FALSE))
+  expect_true(testSubset(NULL, letters, empty.ok = TRUE))
   expect_false(testSubset(factor("a"), letters))
   expect_true(testSubset(1., 1:2))
   expect_true(testSubset(factor("a"), factor(letters)))