+++ NEWS.md
@@ -4,6 +4,7 @@
   All assertions now signal an `assertion_error`.
 * The set family of functions is now more restrict regarding the class, e.g.
   they differentiate between factors and characters.
+* New argument `null.ok` (defaults to `FALSE`) for `*Choice`.
 
 # Version 1.8.2
 * `*Matrix` and `*Array` now additionally allow to check for integerish storage